add log4j2 in gradle - should replace all system.out.println()
