+++ build.gradle
@@ -7,9 +7,6 @@ apply plugin: 'java'
 // code coverage
 apply plugin: 'jacoco'
 
-
-sourceCompatibility = 1.8
-
 repositories {
     mavenCentral()
 }
@@ -30,7 +27,11 @@ repositories {
 }
 
 dependencies {
-    testCompile group: 'junit', name: 'junit', version: '4.11'
+    testCompile group: 'junit', name: 'junit', version: '4.+'
+    compile group: 'org.apache.logging.log4j', name: 'log4j-api', version: '2.6.2'
+    compile group: 'org.apache.logging.log4j', name: 'log4j-core', version: '2.6.2'
+
+    testCompile "org.mockito:mockito-core:1+"
 }
 
 task getHomeDir << {
