+++ src/main/java/com/and1/NRA_Algorithm_Sort.java
@@ -5,7 +5,6 @@ public class NRA_Algorithm_Sort implements Comparable<NRA_Algorithm_Sort>
 	private final int index;
 	private final float distance;
 
-	//Tupel in dem Img und dazugehörige Distance gespeichert wird
 	public NRA_Algorithm_Sort(int index, float distance)
 	{
 		this.index = index;
