+++ src/main/java/com/and1/SimilarityAlgorithm.java
@@ -1,5 +1,7 @@
 package com.and1;
 
+import com.and1.img.MRImage;
+
 import java.util.Vector;
 
 /**
