+++ src/main/java/com/and1/SortL1Distance.java
@@ -1,5 +1,7 @@
 package com.and1;
 
+import com.and1.img.MRImage;
+
 public class SortL1Distance implements Comparable<SortL1Distance>
 {
 	private final float distance;
