+++ src/main/java/com/and1/img/DominantColors.java
@@ -1,4 +1,4 @@
-package com.and1;
+package com.and1.img;
 
 import java.util.Arrays;
