+++ src/main/java/com/and1/img/histogram/HistogramHSI.java
@@ -1,4 +1,4 @@
-package com.and1;
+package com.and1.img.histogram;
 
 import java.awt.*;
 import java.awt.image.BufferedImage;
